server-python/requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.12.0
astroid==3.3.11
attrs==25.4.0
bandit==1.7.9
bcrypt==4.0.1
blinker==1.9.0
boltons==21.0.0
bracex==2.6
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
click-option-group==0.5.9
colorama==0.4.6
defusedxml==0.7.1
dill==0.4.0
distlib==0.4.0
distro==1.9.0
distro-info==1.0
ecdsa==0.19.1
exceptiongroup==1.2.2
face==24.0.0
fastapi==0.112.2
filelock==3.18.0
Flask==3.1.1
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
glom==22.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
openai==2.8.1
packaging==25.0
passlib==1.7.4
peewee==3.18.3
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pylint==3.3.7
pytest==8.3.3
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.4
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.15
ruff==0.5.7
semgrep==1.79.0
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.38.6
stevedore==5.5.0
tomli==2.0.2
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.30.6
virtualenv==20.32.0
wcmatch==8.5.2
Werkzeug==3.1.3
zipp==3.23.0
